refactor: 코드 리팩터링 및 인스턴스 관리 클래스 제거
